requirement.txt
azure-search-documents
azure-core
pandas